IDE_SQUARE_LLM-main/pyproject.toml
[project]
name = "IDE_SQUARE_LLM"
version = "0.1.0"
description = "LLM-driven Multi-Agent System for SQUARE Logic IDE"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=1.0.1",
    "pydantic>=2.12.3",
    "openai>=1.40.0",
    "networkx>=3.2",
    "z3-solver>=4.12.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
optional = [
    "matplotlib>=3.8.0",
    "graphviz>=0.20.0",
    "jupyter>=1.0.0",
]

[project.scripts]
square-ide = "src.main:app"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.ruff]
target-version = "py310"
line-length = 88
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
